CollectivePage: move updates section to GraphQL V2 (#9366)
